requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
werkzeug>=2.3.0
scikit-image>=0.21.0
reportlab>=4.0.0

# Jupyter notebook
notebook>=7.0.0
ipykernel>=6.25.0